fix(nn): the VAE, the conditional UNet and the simple autoencoder norm in their compute dtype
The diffusers-ported VAE encoder and decoder, UNet2DCondition's spatial
attention and output norms and SimpleAutoEncoder's GroupNorms left their
dtype to flax's promotion, which returns fp32 for a bf16 input against fp32
scales, so each bf16 model handed fp32 activations to the swish and the
convolution after every norm. They now take the module's dtype. Parameter
names are unchanged, and fp32 models are unaffected.

test_a_bf16_convolutional_model_keeps_its_activations_in_bf16 extends the
Unet test to the six models: on the previous commit unet_condition,
vae_encoder, vae_decoder, simple_encoder and simple_decoder fail with fp32
norm outputs; all seven cases pass here. The diffusers and SD3 VAE
fixtures (test_vae, test_vae_16ch, test_native_diffusion, test_lora,
test_sd3_run, test_autoencoder) pass unchanged.

+++ src/dew/nn/autoencoders/simple.py
@@ -39,7 +39,8 @@ class SimpleEncoder(nn.Module):
                 name=f"down_{i}",
             )(x)
             if self.norm_groups > 0:
-                x = nn.GroupNorm(_group_count(features, self.norm_groups), name=f"down_norm_{i}")(x)
+                x = nn.GroupNorm(_group_count(features, self.norm_groups), dtype=self.dtype,
+                                 name=f"down_norm_{i}")(x)
             x = self.activation(x)
         return nn.Conv(
             features=self.latent_channels,
@@ -77,7 +78,8 @@ class SimpleDecoder(nn.Module):
                 name=f"up_{i}",
             )(z)
             if self.norm_groups > 0:
-                z = nn.GroupNorm(_group_count(features, self.norm_groups), name=f"up_norm_{i}")(z)
+                z = nn.GroupNorm(_group_count(features, self.norm_groups), dtype=self.dtype,
+                                 name=f"up_norm_{i}")(z)
             z = self.activation(z)
         return nn.Conv(
             features=self.out_channels,
